fix: include destination token in token list (#709)
Fixes an issue created by this feature
https://github.com/hyperlane-xyz/hyperlane-warp-ui-template/pull/701

Removing tokens with no connections would mean that the destination
token could be removed and would trigger the assert here:

https://github.com/hyperlane-xyz/hyperlane-monorepo/blob/288b7799d3cc35bdf60eac632c7015971d6f2021/typescript/sdk/src/warp/WarpCore.ts#L108-L115
And example config would be this, where `sepolia` has no connections but
`basesepolia` still needs the token to connect.
```
tokens:
  - addressOrDenom: '0x5c12ADC734699C07b095fe30B8312F1A7bbaA788'
    chainName: basesepolia
    connections:
      - token: ethereum|sepolia|0x95878Fd41bC26f7045C0b98e381c22f010745A75
    decimals: 18
    name: Ether
    standard: EvmHypNative
    symbol: ETH
  - addressOrDenom: '0x95878Fd41bC26f7045C0b98e381c22f010745A75'
    chainName: sepolia
    decimals: 18
    name: Ether
    standard: EvmHypNative
    symbol: ETH
```

The fix will now include tokens that are destination tokens in the list
and truly only remove those that have no connections nor dependencies
(an example would be some of the chains in `oUSDT/production`)

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

* **Bug Fixes**
* Improved token filtering to ensure tokens that are only destinations
(without outgoing connections) are now correctly displayed.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->